numberOfBiscuit の型を int に修正

+++ YCNo47.java
@@ -3,9 +3,9 @@ import java.util.Scanner;
 public class YCNo47 {
     public static void main(String[] args){
         try(Scanner scan = new Scanner(System.in)){
-            long numberOfBiscuit;
+            int numberOfBiscuit;
             do{
-                numberOfBiscuit = Long.parseLong(scan.nextLine());
+                numberOfBiscuit = Integer.parseInt(scan.nextLine());
             }while(isNotNumberOfBiscuitRange(numberOfBiscuit));
             System.out.println(calcClapNumber(numberOfBiscuit));
         }
@@ -15,7 +15,7 @@ public class YCNo47 {
      * @param number
      * @return
      */
-    private static boolean isNotNumberOfBiscuitRange(long number){
+    private static boolean isNotNumberOfBiscuitRange(int number){
         if(number < 1 || number > Math.pow(10,8) )
             return true;
         return false;
@@ -25,8 +25,8 @@ public class YCNo47 {
      * @param numberOfBiscuit
      * @return
      */
-    private static long calcClapNumber(long numberOfBiscuit){
-        long clapNumber = 0;
+    private static int calcClapNumber(int numberOfBiscuit){
+        int clapNumber = 0;
         while(Math.pow(2,clapNumber) < numberOfBiscuit)
             clapNumber++;
         return clapNumber;
